+++ .env-example
@@ -1,0 +1,4 @@
+REDDIT_CLIENT_SECRET_KEY=
+REDDIT_CLIENT_ID=
+REDDIT_APP_NAME=TestApp
+REDDIT_REDIRECT_URI=ya.ru
