+++ .flake8
@@ -1,0 +1,11 @@
+[flake8]
+ingore = E203, E266, E501, W503, G200, G201, P103
+max-line-length = 88
+max-complexity = 8
+exclude = .git,.env,pg_migrations,migrations,env,venv
+per-file-ignores =
+    __init__.py: F401, ABS101, F403
+    */tests/*: TAE001, FNE006, CFQ002
+    conftest.py: TAE001, FNE008, FNE006
+    **/conftest.py: TAE001, FNE008, FNE006
+extend-ignore = E203
