+++ app/Enums.py
@@ -1,0 +1,6 @@
+import enum
+
+
+class PopularTextTypes(enum.Enum):
+    comment = "комментариев"
+    post = "постов"
