+++ app/redit_app.py
@@ -1,35 +1,43 @@
 import dataclasses
-from typing import Iterator, Callable
 
 import praw
 
-from app.schema import SubbReditSchema
-from config import settings
+from app.config import settings
+from app.mixins import (
+    ConvertSchemasMixin,
+    CountMixins,
+    FilterPostMixin,
+    GetSubbreditsMixin,
+    OtherMixins,
+)
+from app.schemas import RedditCommentSchema, RedditPostSchema, SubbReditSchema
 
 
 @dataclasses.dataclass(slots=True)
-class Reddit:
+class Reddit(ConvertSchemasMixin, FilterPostMixin, GetSubbreditsMixin, CountMixins, OtherMixins):
     reddit: praw.Reddit = praw.Reddit(
         client_id=settings.REDDIT_CLIENT_ID,
         client_secret=settings.REDDIT_CLIENT_SECRET_KEY,
         redirect_uri=settings.REDDIT_REDIRECT_URI,
         user_agent=settings.REDDIT_USER_AGENT,
     )
-    subbredits_list: list[SubbReditSchema] = None
 
     def get_list_of_popular_subreddits(self) -> list[SubbReditSchema]:
         reddits_iterator = self.reddit.subreddits.popular()
-        subbredits_parsed_to_schema = self.parse_subreddits_to_schema(reddits_iterator)
-        self.subbredits_list = subbredits_parsed_to_schema
+        subbredits_parsed_to_schema = self.convert_subreddits_to_schema(reddits_iterator)
         return subbredits_parsed_to_schema
 
-    @staticmethod
-    def parse_subreddits_to_schema(subbredits: Iterator) -> list[SubbReditSchema]:
-        subbredits_list = []
-        for index, subreddit in enumerate(subbredits, 1):
-            full_url = settings.REDDIT_BASE_URL + subreddit.url
-            *_, name_subbredit, _ = subreddit.url.split('/')
-            subbredits_list.append(
-                SubbReditSchema.model_validate({'index': index, 'name': name_subbredit, 'full_url': full_url}))
-        return subbredits_list
+    def grab_hot_posts_from_subbredid(self, subbredid: SubbReditSchema, limit=200) -> list[RedditPostSchema]:
+        top_posts = self.reddit.subreddit(subbredid.name).hot(limit=limit)
+        post_parsed_as_schema = self.convert_posts_to_schema(subbredid.name, top_posts)
+        return post_parsed_as_schema
 
+    def get_comments_from_post(self, post: praw.reddit.Submission, limit: int = 15) -> list[RedditCommentSchema] | None:
+        top_comments = post.comments.list()[:limit]
+        result_list = []
+        for comment in top_comments:
+            parsed_comment = self.convert_comment_to_schema(comment)
+            if parsed_comment:
+                result_list.append(parsed_comment)
+
+        return result_list if result_list else None
