+++ pyproject.toml
@@ -1,0 +1,21 @@
+[tool.mypy]
+ignore_missing_imports = true
+disallow_incomplete_defs = true
+disallow_untyped_calls = true
+disallow_untyped_defs = true
+
+[tool.isort]
+profile = "black"
+multi_line_output = 3
+include_trailing_comma = true
+force_grid_wrap = 0
+use_parentheses = true
+ensure_newline_before_comments = true
+line_length = 88
+
+[tool.autoflake]
+check_diff = true
+imports = ["sqlachemy", "fastapi", "app"]
+
+[tool.pyright]
+include = ["app"]
